+++ src/main/java/net/yxiao233/ifeu/common/utils/KeyDownUtil.java
@@ -1,8 +1,11 @@
 package net.yxiao233.ifeu.common.utils;
 
 import net.minecraft.client.Minecraft;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
 import org.lwjgl.glfw.GLFW;
 
+@OnlyIn(Dist.CLIENT)
 public class KeyDownUtil {
     private static final long window = Minecraft.getInstance().getWindow().getWindow();
     public static boolean isShiftKeyDown(){
@@ -12,14 +15,14 @@ public class KeyDownUtil {
     }
 
     public static boolean isCtrlKeyDown(){
-        boolean isLeftCtrlDown = GLFW.glfwGetKey(window,GLFW.GLFW_KEY_LEFT_CONTROL) == GLFW.GLFW_PRESS;
-        boolean isRightCtrlDown = GLFW.glfwGetKey(window,GLFW.GLFW_KEY_RIGHT_CONTROL) == GLFW.GLFW_PRESS;
+        boolean isLeftCtrlDown = GLFW.glfwGetKey(window, GLFW.GLFW_KEY_LEFT_CONTROL) == GLFW.GLFW_PRESS;
+        boolean isRightCtrlDown = GLFW.glfwGetKey(window, GLFW.GLFW_KEY_RIGHT_CONTROL) == GLFW.GLFW_PRESS;
         return isLeftCtrlDown || isRightCtrlDown;
     }
 
     public static boolean isAltKeyDown(){
-        boolean isLeftAltDown = GLFW.glfwGetKey(window,GLFW.GLFW_KEY_LEFT_ALT) == GLFW.GLFW_PRESS;
-        boolean isRightAltDown = GLFW.glfwGetKey(window,GLFW.GLFW_KEY_RIGHT_ALT) == GLFW.GLFW_PRESS;
+        boolean isLeftAltDown = GLFW.glfwGetKey(window, GLFW.GLFW_KEY_LEFT_ALT) == GLFW.GLFW_PRESS;
+        boolean isRightAltDown = GLFW.glfwGetKey(window, GLFW.GLFW_KEY_RIGHT_ALT) == GLFW.GLFW_PRESS;
         return isLeftAltDown || isRightAltDown;
     }
 }