+++ changelog.md
@@ -92,6 +92,10 @@
 ### v2.0.110
 - 配置工具可以同步配置各种插件了
 
+### v2.1.000
+- 修复了服务端无法打开时间,天气掌控者的bug
+- 修复了服务端无法监听ctrl等按键按下的bug
+
 ## en_us
 ### v1.0.0
 - init
@@ -182,3 +186,7 @@
 
 ### v2.0.110
 - The configuration tool can synchronize the configuration of various addons now
+
+### v2.1.000
+- Fixed the bug where the server cannot open the time and weather controller
+- Fixed the bug where the server could not listen for keystrokes such as Ctrl